+++ core/src/main/java/com/cobber/fta/KeywordEntry.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2017-2022 Tim Segall
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.cobber.fta;
+
+/**
+ * Each KeywordEntry provided a pair with the locale and the value of the tag in that locale.
+ * See keywords.json for the structure of the JSON file.
+ */
+public class KeywordEntry {
+	public String localeTag;
+	public String value;
+}